:art: Improve decision schedule #37
