+++ Sources/Application/Extensions/ObservableType+.swift
