+++ Sources/Application/Helper/ReinforcementResult.swift
